+++ next-docs/public/examples/select/singleReactSelect/WithLabel.tsx
@@ -21,14 +21,14 @@ const Example = () => (
     <div className="w-80 mr-4 ml-4 mb-8">
       <Select
         options={options}
-        label="Single md"
+        label="Large"
         placeholderSlot="Choose an option"
       />
     </div>
     <div className="w-80 mr-4 ml-4 mb-8">
       <Select
         options={options}
-        label="Single lg"
+        label="xLarge"
         size="xLarge"
         placeholderSlot="Choose an option"
       />