+++ packages/table/src/utils/renderSpanRows.tsx
@@ -3,12 +3,15 @@ import BodyTR from '../components/BodyTR';
 import TD from '../components/TD';
 import { RowSpanHeader } from '../hooks/useRowSpan';
 import { ColorNames } from '@heathmont/moon-themes';
+import { Cell, Row } from 'react-table';
 
-type RenderSpanRowsProps = {
-  rows: any;
-  prepareRow: any;
+type RenderSpanRowsProps<D extends object = {}> = {
+  rows: Row<D>[];
+  prepareRow: (row: Row<D>) => void;
   rowSpanHeaders: RowSpanHeader[];
-  getOnRowClickHandler: (row: any) => any;
+  getOnRowClickHandler: (
+    row: Row<D>
+  ) => ((row: Row<D>) => void | (() => void)) | undefined;
   evenRowBackgroundColor: ColorNames;
   defaultRowBackgroundColor: ColorNames;
 };
@@ -22,29 +25,30 @@ const renderSpanRows = ({
   rowSpanHeaders,
 }: RenderSpanRowsProps) => {
   if (!rows) return;
-  return rows.map((row: any, index: number) => {
+  return rows.map((row: Row<{}>, index: number) => {
     prepareRow(row);
 
-    const onRowClickHandler = getOnRowClickHandler(row);
-    const hasOnRowClickHandler = typeof onRowClickHandler === 'function';
+    const onRowClickHandler = getOnRowClickHandler
+      ? getOnRowClickHandler(row)
+      : () => undefined;
     const backgroundColor =
       index % 2 ? evenRowBackgroundColor : defaultRowBackgroundColor;
     const isRowSpanned =
       rowSpanHeaders &&
-      rowSpanHeaders.some((rowSpanHeader: any) =>
+      rowSpanHeaders.some((rowSpanHeader: RowSpanHeader) =>
         row.cells.some(
-          (cell: any) =>
+          (cell: Cell<{}>) =>
             cell.column &&
             rowSpanHeader.id === cell.column.id &&
             rowSpanHeader.value === cell.value
         )
       );
 
-    const makeCellForRowSpanned = (cell: any) => (
+    const makeCellForRowSpanned = (cell: Cell<{}>) => (
       <TD {...cell.getCellProps()} />
     );
 
-    const makeCellForNormalRow = (cell: any) => (
+    const makeCellForNormalRow = (cell: Cell<{}>) => (
       <TD {...cell.getCellProps()}>{cell.render('Cell')}</TD>
     );
 
@@ -53,11 +57,9 @@ const renderSpanRows = ({
         {...row.getRowProps()}
         withOffset={!isRowSpanned}
         backgroundColor={backgroundColor}
-        onClick={
-          hasOnRowClickHandler ? () => onRowClickHandler(row) : undefined
-        }
+        onClick={onRowClickHandler ? () => onRowClickHandler(row) : undefined}
       >
-        {row.cells.map((cell: any) => {
+        {row.cells.map((cell: Cell<{}>) => {
           if (!rowSpanHeaders) return makeCellForNormalRow(cell);
 
           const rowSpanHeader = rowSpanHeaders.find(