+++ packages/select/CHANGELOG.md
@@ -1,6 +1,22 @@
 # @heathmont/moon-select
 
+## 7.21.0
+
+### Minor Changes
+
+- New nav & select fixes
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-core@7.21.0
+  - @heathmont/moon-components@7.21.0
+  - @heathmont/moon-icons@7.21.0
+  - @heathmont/moon-themes@7.21.0
+  - @heathmont/moon-utils@7.21.0
+
 ## 7.20.0
+
 ### Minor Changes
 
 - add moon-select package