+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 7.21.0
+
+### Minor Changes
+
+- New nav & select fixes
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-themes@7.21.0
+
 ## 7.20.0
 
 ### Minor Changes